SysML modeling
