--- cloudmodel/go.mod
@@ -1,5 +1,0 @@
-module github.com/sdoque/systems/cloudmodel
-
-go 1.25.8
-
-require github.com/sdoque/mbaigo v0.1.0-alpha.3
