+++ modeler/go.mod
@@ -1,0 +1,5 @@
+module github.com/sdoque/systems/modeler
+
+go 1.25.8
+
+require github.com/sdoque/mbaigo v0.1.0-alpha.3
